[IT-4225] Clarify image version
There are cases where you may want to update the ImageVersion number
without needing to update the underlying template.  For example,
when you want to update a passed in parameter but keep the template
the same.  In those situation the ImageVersion still needs to be
incremented.  The convention of setting the ImageVersion number to
the aws-infra template does not work for that use case.  Update the
convention to just use an arbitrary version number.

Also start over with agora-data-manager-github-runner image